# server/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VPS CI/CD Automation Runner API",
    "description": "Webhook-driven git sync, automated zero-downtime deployment runner, and telemetry platform for self-hosted VPS instances.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
  "paths": {
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get authenticated user information",
        "responses": {
          "200": {
            "description": "Current user status"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticate administrator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign in successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List all configured sync services",
        "responses": {
          "200": {
            "description": "List of services"
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Create a new sync service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/triggers": {
      "get": {
        "tags": [
          "Triggers"
        ],
        "summary": "List recent triggers across all services",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 40
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "service_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trigger list"
          }
        }
      }
    },
...
    "/api/auth/security-question": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get the configured security question",
        "responses": {
          "200": {
            "description": "Question state"
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Set or update the security question and answer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string"
                  },
                  "answer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset the password by answering the security question",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "answer",
                  "newPassword"
                ],
                "properties": {
                  "answer": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset & signed in"
          },
          "401": {
            "description": "Incorrect answer"
          },
          "404": {
            "description": "No security question set"
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "Get server settings",
        "responses": {
          "200": {
            "description": "Settings"
          }
        }
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "Update the public base URL setting",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "publicBaseUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved"
          }
        }
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change the administrator password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string",
                    "minLength": 8
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Current password incorrect"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ServiceInput": {
        "type": "object",
        "required": [
          "name",
          "repo_url",
          "folder_path"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "github",
              "gitlab",
              "bitbucket",
              "gitea",
              "gogs",
              "generic"
            ]
          },
          "repo_url": {
            "type": "string"
          },
          "folder_path": {
            "type": "string"
          },
          "branch_mode": {
            "type": "string",
            "enum": [
              "webhook",
              "fixed",
              "current"
            ]
          },
          "fixed_branch": {
            "type": "string"
          },
          "allowed_branches": {
            "type": "string"
          },
          "sync_mode": {
            "type": "string",
            "enum": [
              "pull",
              "reset"
            ]
          },
          "healthcheck_url": {
            "type": "string"
          },
          "auto_rollback": {
            "type": "boolean"
          },
          "commands": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "command": {
                  "type": "string"
                },
                "branch_filter": {
                  "type": "string"
                },
                "continue_on_error": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  }
}